requirements.txt
# PyTorch: plain `pip install -r` often installs CPU-only wheels.
# For NVIDIA GPU, install CUDA build first, e.g.:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
PyYAML>=6.0
matplotlib>=3.7.0
